% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download Canadian Premier League Centre Circle Data}
\usage{
download_data(email, url, dir)
}
\arguments{
\item{email}{E-mail associated with Google account.}

\item{url}{Link to the Centre Circle Data Google Drive folder.}

\item{dir}{Folder in which the data will be saved.}
}
\value{
CSV files saved in given directory.
}
\description{
Download Canadian Premier League Centre Circle Data
}
\examples{
\dontrun{
download_data(
  "first.last@example.com",
  "https://drive.google.com/drive/folders/abc123",
  "~/cpl_data/"
)
}
}
