% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read Canadian Premier League Centre Circle Data}
\usage{
read_data(dir)
}
\arguments{
\item{dir}{Directory containing the data.}
}
\value{
A list of data frames.
}
\description{
Read Canadian Premier League Centre Circle Data
}
\examples{
\dontrun{
cpl_data <- read_data("~/cpl_data/")
}
}
