% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{update_data}
\alias{update_data}
\title{Update Canadian Premier League Centre Circle Data}
\usage{
update_data(email, url, dir)
}
\arguments{
\item{email}{E-mail associated with Google account.}

\item{url}{Link to the Centre Circle Data Google Drive folder.}

\item{dir}{Folder containing the data.}
}
\value{
CSV files saved in given directory.
}
\description{
Update Canadian Premier League Centre Circle Data
}
\examples{
#'
\dontrun{
update_data(
  "first.last@example.com",
  "https://drive.google.com/drive/folders/abc123",
  "~/cpl_data/"
)
}
}
